% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{add_mip_sol}
\alias{add_mip_sol}
\title{Add MIP Solution}
\usage{
add_mip_sol(prob, heur_sol)
}
\arguments{
\item{prob}{the Xpress problem object}

\item{heur_sol}{A dataframe containing column variable names, and solutions}
}
\description{
Add MIP Solution
}
