% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{build_row_equation}
\alias{build_row_equation}
\title{Build Row Constraint Equation}
\usage{
build_row_equation(prob, row)
}
\arguments{
\item{prob}{the Xpress problem object}

\item{row}{an integer value of row value}
}
\description{
Build Row Constraint Equation
}
