% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{build_row_equation_from_parts}
\alias{build_row_equation_from_parts}
\title{Build Row Constraint Equation From Parts}
\usage{
build_row_equation_from_parts(row_data, row_col_names, row_types, row_rhs)
}
\arguments{
\item{row_data}{The Xpress problem row information from xpress::getrows for a single row}

\item{row_col_names}{A vector of all Xpress problem column names from xpress::getnamelist (will be filtered in function)}

\item{row_types}{The Xpress problem row type from xpress::getrowtype that has been convereted
from alpha-numeric to math symbols for a single row}

\item{row_rhs}{The Xpress problem row right-hand side from xpress::getrhs for a single row}
}
\description{
This function constructs the row constraint equation from passing in all the various parts.
}
