% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{get_ompr_status}
\alias{get_ompr_status}
\title{Get OMPR Optimized Model Status}
\usage{
get_ompr_status(x)
}
\arguments{
\item{x}{an integer value representing the staut of the optimized model}
}
\description{
Get OMPR Optimized Model Status
}
