% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{get_row_equations}
\alias{get_row_equations}
\title{Get Row Constraint Equations}
\usage{
get_row_equations(prob, rows)
}
\arguments{
\item{prob}{the Xpress problem object}

\item{row}{an integer vector of rows to retreive data for.}
}
\description{
Get Row Constraint Equations
}
