% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{get_xpress_status}
\alias{get_xpress_status}
\title{Get Xpress Optimized Model Status}
\usage{
get_xpress_status(x)
}
\arguments{
\item{x}{an integer value representing the status of the optimized model}
}
\description{
Get Xpress Optimized Model Status
}
