% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{run_inf_analysis}
\alias{run_inf_analysis}
\title{Run Xpress Infeasibility Analysis}
\usage{
run_inf_analysis(prob)
}
\arguments{
\item{prob}{the Xpress problem object}
}
\description{
Run Xpress Infeasibility Analysis
}
