% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{xpress_apply_control_params}
\alias{xpress_apply_control_params}
\title{Apply Xpress Control Parameters}
\usage{
xpress_apply_control_params(prob, control_params)
}
\arguments{
\item{prob}{the Xpress problem object}

\item{control_params}{the list of control paramters to pass to the solver}
}
\description{
Apply Xpress Control Parameters
}
