% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{xpress_col_type_conversion}
\alias{xpress_col_type_conversion}
\title{OMPR Column Type Conversion}
\usage{
xpress_col_type_conversion(col_type)
}
\arguments{
\item{col_type}{a string representing the column type}
}
\description{
OMPR Column Type Conversion
}
