% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{xpress_col_type_fill}
\alias{xpress_col_type_fill}
\title{Xpress Column Type Fill}
\usage{
xpress_col_type_fill(x, conversion_list)
}
\arguments{
\item{x}{a string representing the column type description}

\item{conversion_list}{a list of the conversion parameters}
}
\description{
Xpress Column Type Fill
}
