% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{xpress_convert_row}
\alias{xpress_convert_row}
\title{Convert Row From Alpha-Numeric to Symbol}
\usage{
xpress_convert_row(row_type)
}
\arguments{
\item{sense}{the Alpha-Numeric Row Type}
}
\description{
Convert Row From Alpha-Numeric to Symbol
}
