% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{xpress_get_params}
\alias{xpress_get_params}
\title{Get Xpress Problem Parameters}
\usage{
xpress_get_params(prob)
}
\arguments{
\item{prob}{the Xpress problem object}
}
\description{
Get Xpress Problem Parameters
}
