% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{xpress_sense_conversion}
\alias{xpress_sense_conversion}
\title{Get Row Type Conversion}
\usage{
xpress_sense_conversion(sense)
}
\arguments{
\item{sense}{the math symbol representing the row constraint type}
}
\description{
Get Row Type Conversion
}
