% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpress_optimizer.R
\name{xpress_optimizer}
\alias{xpress_optimizer}
\title{Run the Xpress solver through the OMPR interface.}
\usage{
xpress_optimizer(control = list(problem_name = "Xpress Problem"), ...)
}
\arguments{
\item{control}{a list of options passed to \code{xpress::xprs_optimize()}.
A complete list of control parameters are availalbe in the details section.}
}
\value{
A list returning all the results and parameters from the Xpress optimization run.
}
\description{
Run the Xpress solver through the OMPR interface.
}
\details{
The Xpress solver offers an extraordinarily wide range of control settings. The full list of
controls available can be found within the links below. While the controls are broken out
into integer, double, and string controls, there's no need to delineate this in your function
call as that will be handled internally.

General Control Parameter Discussion With List of Controls
https://www.fico.com/fico-xpress-optimization/docs/latest/solver/optimizer/HTML/chapter7.html

Integer Control Parameters
https://www.fico.com/fico-xpress-optimization/docs/latest/solver/optimizer/HTML/XPRSsetintcontrol.html#'

Double Control Parameters
https://www.fico.com/fico-xpress-optimization/docs/latest/solver/optimizer/HTML/XPRSsetdblcontrol.html

String Control Parameters
https://www.fico.com/fico-xpress-optimization/docs/latest/solver/optimizer/HTML/XPRSsetstrcontrol.html
}
\examples{
\dontrun{



}
}
